Divided index.js codes to modules
