+++ index.html
@@ -23,7 +23,7 @@
           </li>
         </ul>
       </nav>
-      <div id="time">Febuary 9th 2022, 1:43pm</div>
+      <div id="time"></div>
     </header>
     <section class="book-list">
       <h1>All awesome books</h1>
@@ -70,8 +70,13 @@
     <footer>
       <p class="copyright">Copyright</p>
     </footer>
-    <script src="./modules/app.js" type="module"></script>
+    <script src="./modules/data.js" type="module"></script>
     <script src="./modules/book.js" type="module"></script>
+    <script src="./modules/addBooksListener.js" type="module"></script>
+    <script src="./modules/addNavListeners.js" type="module"></script>
+    <script src="./modules/checkAndDisplayList.js" type="module"></script>
+    <script src="./modules/displayAllBooksListener.js" type="module"></script>
+    <script src="./modules/domElements.js" type="module"></script>
     <script src="luxon.js"></script>
     <script src="index.js" type="module"></script>
   </body>
